requirements.txt
# requirements.txt

Flask
openai
python-dotenv
streamlit